code/requirements.txt
openai==0.28
pillow
langchain==0.0.304
langchain_community
boto3
opensearch-py==2.3.1
retrying==1.3.4
botocore
apache-beam==2.50.0
datasets==2.14.5
tiktoken==0.4.0
pydantic==1.10.13
pypdf==4.1.0
faiss-cpu==1.8.0
sqlalchemy==2.0.28
pinecone-client
python-dotenv
requests_aws4auth
streamlit
streamlit_pdf_viewer